Refactored additional pregen dimensions config
Replaced the custom BaseValue implementation for additional pregen dimensions with a dedicated ResourceLocationListValue class.
